requirements.txt
discord
python-dotenv
python-decouple
finnhub-python
websocket_client